LayoutTests/http/wpt/webauthn/public-key-credential-create-failure-nfc.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[create]] with timeout in a mock nfc authenticator.
PASS PublicKeyCredential's [[create]] with no tags in a mock nfc authenticator.
PASS PublicKeyCredential's [[create]] with wrong tag type in a mock nfc authenticator.
PASS PublicKeyCredential's [[create]] with no connections in a mock nfc authenticator.
PASS PublicKeyCredential's [[create]] with null version in a mock nfc authenticator.
PASS PublicKeyCredential's [[create]] with wrong version in a mock nfc authenticator.
PASS PublicKeyCredential's [[create]] with wrong version in a mock nfc authenticator.2
